Move preset test to new package
