+++ backend/src/test/java/com/benine/backend/preset/PresetControllerTest.java
@@ -1,5 +1,6 @@
-package com.benine.backend;
+package com.benine.backend.preset;
 
+import com.benine.backend.ServerController;
 import com.benine.backend.database.MySQLDatabase;
 import com.benine.backend.preset.Preset;
 import com.benine.backend.preset.PresetController;
