+++ backend/src/test/java/com/benine/backend/preset/PresetTest.java
@@ -1,4 +1,4 @@
-package com.benine.backend;
+package com.benine.backend.preset;
 
 import org.json.JSONException;
 import org.json.simple.JSONArray;
